Initial Upload
Worked on this project throughout the first couple of months at VPI. When starting this project, I had a limited understanding of HTML and CSS, barely any understanding of JavaScript (I knew it wasn't Java..), no understanding of the DOM, I didn't know what a CDN was, I didn't understand that frameworks were adaptions of the original language (JQuery, Angular, React, etc.). This project is a demonstration of my growth in these areas.